+++ Libplanet/Store/StoreExtension.cs
@@ -1,6 +1,7 @@
 using System;
-using System.Collections.Generic;
+#pragma warning disable S1128 // Remove this unnecessary 'using'
 using System.Collections.Immutable;
+#pragma warning restore S1128 // Remove this unnecessary 'using'
 using System.Linq;
 using System.Security.Cryptography;
 using Libplanet.Action;
@@ -42,58 +43,6 @@ namespace Libplanet.Store
 
             return store.IterateStateReferences(chainId, address, lookupUntil.Index, limit: 1)
                     .FirstOrDefault();
-        }
-
-        /// <summary>
-        /// Lists all accounts, that have any states, in the given <paramref name="chainId"/> and
-        /// their state references.
-        /// </summary>
-        /// <param name="store">A store object.</param>
-        /// <param name="chainId">The chain ID to look up state references.</param>
-        /// <param name="onlyAfter">Includes state references only made after the block
-        /// this argument refers to, if present.</param>
-        /// <param name="ignoreAfter">Excludes state references made after the block
-        /// this argument refers to, if present.</param>
-        /// <returns>A dictionary of account addresses to lists of their corresponding state
-        /// references.  Each list of state references is in ascending order, i.e., the block
-        /// closest to the genesis goes first and the block closest to the tip goes last.</returns>
-        public static IImmutableDictionary<Address, IImmutableList<HashDigest<SHA256>>>
-        ListAllStateReferences(
-            this IStore store,
-            Guid chainId,
-            HashDigest<SHA256>? onlyAfter = null,
-            HashDigest<SHA256>? ignoreAfter = null
-        )
-        {
-            (HashDigest<SHA256>, long)? baseBlock =
-                onlyAfter is HashDigest<SHA256> @base && store.GetBlockIndex(@base) is long baseIdx
-                    ? (@base, baseIdx)
-                    : null as (HashDigest<SHA256>, long)?;
-            (HashDigest<SHA256>, long)? targetBlock =
-                ignoreAfter is HashDigest<SHA256> tgt && store.GetBlockIndex(tgt) is long tgtIdx
-                    ? (tgt, tgtIdx)
-                    : null as (HashDigest<SHA256>, long)?;
-            return store.ListAddresses(chainId).Select(address =>
-            {
-                IEnumerable<Tuple<HashDigest<SHA256>, long>> refIndices =
-                    store.IterateStateReferences(chainId, address);
-
-                if (targetBlock is ValueTuple<HashDigest<SHA256>, long> targetIndex)
-                {
-                    refIndices = refIndices.SkipWhile(p => p.Item2 > targetIndex.Item2);
-                }
-
-                if (baseBlock is ValueTuple<HashDigest<SHA256>, long> baseIndex)
-                {
-                    refIndices = refIndices.TakeWhile(p => p.Item2 > baseIndex.Item2);
-                }
-
-                ImmutableList<HashDigest<SHA256>> refs = refIndices
-                    .Select(p => p.Item1)
-                    .Reverse()
-                    .ToImmutableList();
-                return new KeyValuePair<Address, IImmutableList<HashDigest<SHA256>>>(address, refs);
-            }).Where(pair => pair.Value.Any()).ToImmutableDictionary();
         }
     }
 }